Update de-DE.json
This commit fixes grammar and spelling mistakes and improves the tone and style of the german translation.

I tried to be as close as possible to the original text but took some liberties if the sentences structure did not translate well into german.

I also capitalized "Du" (you) to make the text sound slighlty more formal. This is somewhat controversial in german but so is *:innen and other gender specific language and this was here before I touched the file. So I think it's fine :)